+++ pipe.h
@@ -4,9 +4,10 @@
 #include "sync.h"
 #include "types.h"
 #include "config.h"
+#include <stdint.h>
 
 typedef struct pipe {
-    uint8_t *pipe_msg;      //PONTEIRO
+    uint16_t *pipe_msg;      //PONTEIRO
     uint8_t pipe_pos_read;
     uint8_t pipe_pos_write;
     sem_t pipe_sem_read;
@@ -15,8 +16,8 @@ typedef struct pipe {
 
 void create_pipe(pipe_t *p);
 void destroy_pipe(pipe_t *p);
-void write_pipe(pipe_t *p, uint8_t data);
-void read_pipe(pipe_t *p, uint8_t *data);
+void write_pipe(pipe_t *p, uint16_t data);
+void read_pipe(pipe_t *p, uint16_t *data);
 
 #endif	/* PIPE_H */
